US3112 Check off item on list
Now when you click on an item on the list, it will toggle the checked state between true and false in the database record.
If the list item has checked: true, then a CSS class is added which adds a strikethrough. (NOTE: it would be ideal if the strikethrough went right along the line, but this isn't straightforward. TODO!)

Previously, the checked field in the items table was nullable. I've created two migrations:
1. Changing checked from nil to false in all existing item records
2. Making the name, quantity, and checked fields all not nullable, and defaulting checked to false.