US3114 Add a category to the list item
The list page now includes an HTML <select> including all of the possible categories.

The categories are stored in the categories DB table, which is linked to the items table on the category_id field.
The categories DB table is populated from the seed file.

I set this up using several migrations:
• Creating the Categories DB table
• Changing item.category → item.category_id
• Changing item.category_id to a type of bigint (to match the primary key for the category table)
• Linking Items and Categories via a foreign key

Plus I added a belongs_to relation and validation on the Items model.