US3129 Make List.name non-nullable; backfill existing nils.
Add feature test covering the scenario of a user trying to update the name to something invalid.

When editing the list's name, autofocus the <input>.
This is done via JavaScript: if passing autofocus: true, this causes an error on the client-side due the the form element being hidden on page load.

Delete dead tests, update others.