Delete migration: backfilling list (shopping list) names
Normally, I wouldn't do this. The problem here is that List was renamed to ShoppingList: database table, class, files, etc.
So in this migration, there's an error: the List class doesn't exist. However, if we change the class to ShoppingList, that's also a problem, because at this point in time the database table was still called lists, not shopping_lists.

Since this is a small project without people working on it yet, Eddi and I discussed it and decided that it was okay to delete this migration